Prep to replace layerModel classes with ItemTree-based layer stack model
- Factor out a cleaner LayerStackBaseModel class
- Move edit target model, dialog, and actions into outliner module
